@@ -6,26 +6,26 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.4.4</version>
+        <version>2.6.13</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
 
-    <groupId>com.github.ioridazo</groupId>
+    <groupId>github.com.ioridazo</groupId>
     <artifactId>fundanalyzer-patch</artifactId>
-    <version>0.0.3-SNAPSHOT</version>
+    <version>0.0.4</version>
     <name>fundanalyzer-patch</name>
     <description>patch for fundanalyzer application</description>
 
     <properties>
-        <java.version>11</java.version>
-        <spring-cloud.version>2020.0.2</spring-cloud.version>
-        <ecs-logging-java.version>1.0.1</ecs-logging-java.version>
-        <maven-assembly-plugin.version>3.3.0</maven-assembly-plugin.version>
-        <maven-pmd-plugin.version>3.14.0</maven-pmd-plugin.version>
-        <jacoco-maven-plugin.version>0.8.6</jacoco-maven-plugin.version>
-        <spotbugs-maven-plugin.version>4.2.0</spotbugs-maven-plugin.version>
-        <spotbugs.version>4.2.2</spotbugs.version>
+        <java.version>17</java.version>
+        <ecs-logging-java.version>1.5.0</ecs-logging-java.version>
+        <spring-cloud.version>2021.0.4</spring-cloud.version>
 
+        <maven-assembly-plugin.version>3.4.2</maven-assembly-plugin.version>
+        <maven-pmd-plugin.version>3.19.0</maven-pmd-plugin.version>
+        <jacoco-maven-plugin.version>0.8.8</jacoco-maven-plugin.version>
+        <spotbugs-maven-plugin.version>4.7.2.1</spotbugs-maven-plugin.version>
+        <spotbugs.version>4.7.3</spotbugs.version>
     </properties>
 
     <dependencies>
@@ -47,10 +47,6 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-log4j2</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-actuator</artifactId>
         </dependency>
 
         <!-- database -->
@@ -114,6 +110,17 @@
             </dependency>
         </dependencies>
     </dependencyManagement>
+
+    <distributionManagement>
+        <repository>
+            <id>releases</id>
+            <url>http://localhost:8081/repository/maven-releases/</url>
+        </repository>
+        <snapshotRepository>
+            <id>snapshots</id>
+            <url>http://localhost:8081/repository/maven-snapshots/</url>
+        </snapshotRepository>
+    </distributionManagement>
 
     <build>
         <plugins>
